+++ game/game.rb
@@ -1,0 +1,17 @@
+require 'json'
+require 'date'
+require_relative '../item'
+
+class Game < Item
+  attr_accessor :multiplayer, :last_played_at, :publish_date
+
+  def initialize(multiplayer, last_played_at, publish_date, id = nil, archived: false)
+    super(publish_date, id, archived: archived)
+    @multiplayer = multiplayer
+    @last_played_at = Date.parse(last_played_at)
+  end
+
+  def can_be_archived?
+    super && ((Date.today.year - @last_played_at.year) > 2)
+  end
+end
