+++ label/label.rb
