+++ music_album/music_album.rb
@@ -1,0 +1,15 @@
+require 'json'
+require_relative '../item'
+
+class MusicAlbum < Item
+  attr_accessor :on_spotify
+
+  def initialize(on_spotify, publish_date, id = nil, archived: false)
+    super(publish_date, id, archived: archived)
+    @on_spotify = on_spotify
+  end
+
+  def can_be_archived?
+    super && @on_spotify == true
+  end
+end
