+++ schema.sql
@@ -3,26 +3,17 @@ CREATE DATABASE my_catalog;
 CREATE TABLE author(
   ID SERIAL PRIMARY KEY,
   first_name VARCHAR(30),
-  last_name VARCHAR(30),
+  last_name VARCHAR(30)
 );
-
-CREATE TABLE game(
-  ID SERIAL PRIMARY KEY,
-  multiplayer VARCHAR(30) NOT NULL,
-  last_played_at DATE NOT NULL,
-  publish_date DATE NOT NULL,
-  archived BOOLEAN NOT NULL,
-  label_ID INT,
-  author_ID INT,
-  genre_ID INT,
-  FOREIGN KEY (label_ID) REFERENCES label(ID),
-  FOREIGN KEY (author_ID) REFERENCES author(ID),
-  FOREIGN KEY(genre_ID) REFERENCES genre(ID)
-)
   
 CREATE TABLE label(
   ID SERIAL PRIMARY KEY,
-  name VARCHAR(20)
+  name VARCHAR(30)
+);
+
+CREATE TABLE genre(
+  ID SERIAL PRIMARY KEY,
+  name VARCHAR(30)
 );
 
 CREATE TABLE book(
@@ -37,4 +28,28 @@ CREATE TABLE book(
   FOREIGN KEY (label_ID) REFERENCES label(ID),
   FOREIGN KEY (author_ID) REFERENCES author(ID),
   FOREIGN KEY(genre_ID) REFERENCES genre(ID)
-)
+);
+
+CREATE TABLE game(
+  ID SERIAL PRIMARY KEY,
+  multiplayer VARCHAR(30) NOT NULL,
+  last_played_at DATE NOT NULL,
+  publish_date DATE NOT NULL,
+  archived BOOLEAN NOT NULL,
+  label_ID INT,
+  author_ID INT,
+  genre_ID INT,
+  FOREIGN KEY (label_ID) REFERENCES label(ID),
+  FOREIGN KEY (author_ID) REFERENCES author(ID),
+  FOREIGN KEY(genre_ID) REFERENCES genre(ID)
+);
+
+CREATE TABLE music_album(
+  ID SERIAL PRIMARY KEY,
+  publish_date DATE NOT NULL,
+  on_spotify BOOLEAN NOT NULL,
+  archived BOOLEAN NOT NULL,
+  label_ID INT REFERENCES label(ID),
+  author_ID INT REFERENCES author(ID),
+  genre_ID INT REFERENCES genre(ID)
+);
