+++ spec/game_spec.rb
@@ -1,25 +1,27 @@
-require_relative '../game'
+require_relative '../game/game'
+require_relative '../game/game_controller'
 
 describe Game do
+  include GamesController
   it 'Should be an instance of Game' do
-    author = Game.new('Jordan', '04-12-2011', '09-10-2017')
+    author = Game.new(true, '04-12-2011', '09-10-2017')
     expect(author).to be_instance_of Game
   end
 
   it 'archived should be true' do
-    game = Game.new('Jordan', '04-12-2011', '09-10-2017')
+    game = Game.new(true, '04-12-2017', '09-10-2010')
     game.move_to_archive
     game.archived.should eq true
   end
 
   it 'archived should be true' do
-    game = Game.new('Jordan', '04-12-2011', '09-10-2017')
+    game = Game.new(true, '04-12-2011', '09-10-2022')
     game.archived.should eq false
   end
 
   it 'Should store game in json file' do
-    game = Game.new('Jordan', '04-12-2011', '09-10-2017')
-    game.add_game
-    expect(File.exist?('games.json') && File.read('games.json') != '').to eq true
+    game = Game.new(true, '04-12-2011', '09-10-2017')
+    save_game(game)
+    expect(File.exist?('./game/games.json') && File.read('./game/games.json') != '').to eq true
   end
 end
